# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CampusCoin API
  version: 0.1.0
  description: CampusCoin ("Smart Spending, Student Style") REST API. Money amounts are decimal strings (never floats); errors follow RFC 9457 `application/problem+json`; every route scoped to the caller's own data returns 404 (not 403) for another user's resource.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token (EdDSA JWT) from `POST /auth/login`, sent as `Authorization: Bearer <token>`."
  schemas:
...
    copyPreviousBudgetsSchema:
      type: object
      properties:
        month:
          type: string
          pattern: ^(\d{4})-(\d{2})-(\d{2})$
      required:
        - month
      additionalProperties: false
...
    createCategorySchema:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum:
            - income
            - expense
        icon:
          type: string
          pattern: ^[a-z0-9-]{1,40}$
        color:
          type: string
          pattern: ^#[0-9A-Fa-f]{6}$
      required:
        - name
        - type
      additionalProperties: false
...
    updateCategorySchema:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type:
            - string
            - "null"
          pattern: ^[a-z0-9-]{1,40}$
        color:
          type:
            - string
            - "null"
          pattern: ^#[0-9A-Fa-f]{6}$
        isActive:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
          maximum: 32767
      additionalProperties: false
...
    upsertBudgetsSchema:
      type: object
      properties:
        month:
          type: string
          pattern: ^(\d{4})-(\d{2})-(\d{2})$
        budgets:
          type: array
          items:
            type: object
            properties:
              categoryId:
                type: integer
                exclusiveMinimum: 0
              limitAmount:
                type: string
                pattern: ^\d{1,12}(\.\d{1,2})?$
              alertThresholdPct:
                type: integer
                minimum: 50
                maximum: 100
                default: 80
            required:
              - categoryId
              - limitAmount
            additionalProperties: false
          minItems: 1
          maxItems: 200
      required:
        - month
        - budgets
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          description: Machine-readable error slug, e.g. "validation-failed".
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
            required:
              - field
              - message
          description: Present on 422 validation-failed responses, one entry per invalid field.
      required:
        - type
        - title
        - status
      description: RFC 9457 application/problem+json body.
...
paths:
...
  /api/v1/categories:
    get:
      tags:
        - Categories
      summary: List categories (system + own).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            enum:
              - income
              - expense
          required: false
          name: type
          in: query
        - schema:
            type: string
            enum:
              - "true"
              - "false"
          required: false
          name: includeInactive
          in: query
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    post:
      tags:
        - Categories
      summary: Create a custom category.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createCategorySchema"
      responses:
        "201":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/categories/{id}:
    patch:
      tags:
        - Categories
      summary: Rename/recolor a custom category.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^\d{1,10}$
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateCategorySchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      tags:
        - Categories
      summary: Delete (or archive) a custom category.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^\d{1,10}$
          required: true
          name: id
          in: path
        - schema:
            type: string
            pattern: ^\d{1,10}$
          required: false
          name: reassignTo
          in: query
        - schema:
            type: string
            enum:
              - "true"
              - "false"
          required: false
          name: archive
          in: query
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /api/v1/budgets:
    get:
      tags:
        - Budgets
      summary: List budgets for a month.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^(\d{4})-(\d{2})-(\d{2})$
          required: false
          name: month
          in: query
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    put:
      tags:
        - Budgets
      summary: Bulk upsert budgets for a month. Supports `Idempotency-Key`.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/upsertBudgetsSchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/budgets/copy-previous:
    post:
      tags:
        - Budgets
      summary: Copy last month's budgets into the given month.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/copyPreviousBudgetsSchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/budgets/{id}:
    delete:
      tags:
        - Budgets
      summary: Delete a budget line.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^\d{1,10}$
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...